feat: refactored simplified types and added next many command
