+++ runtime/lib/Handlers/function/lib/execute-function/lib/validate-next.ts
@@ -1,8 +1,8 @@
-import { isNextPath, NextCommand } from '../../../runtime-command/next-command.dto';
+import { NextCommand, NextPathDTO } from '../../../runtime-command/next-command.dto';
 import { FunctionPathException } from '../exceptions/function-path.exception';
 
 export function validateNext(next: NextCommand, expectedPathCodes: Array<string>) {
-  if (isNextPath(next) && !expectedPathCodes.includes(next.path)) {
+  if (NextPathDTO.safeParse(next).success && !expectedPathCodes.includes(next.path)) {
     throw new FunctionPathException(next.path, expectedPathCodes);
   }
 }
