+++ runtime/lib/Handlers/function/lib/execute-function/lib/validate-trace.ts
@@ -1,6 +1,7 @@
 import { InvalidRuntimeCommandException } from '@/runtime/lib/HTTPClient/function-lambda/exceptions/invalid-runtime-command.exception';
 
-import { SimpleTraceDTO, Trace } from '../../../runtime-command/trace-command.dto';
+import { Trace } from '../../../runtime-command/trace/base.dto';
+import { SimpleTraceDTO } from '../../../runtime-command/trace-command.dto';
 import { isSimpleTraceType } from './is-simple-trace';
 
 export function parseTrace(trace: Trace): Trace {
