+++ runtime/lib/Handlers/function/runtime-command/constants.ts
@@ -1,0 +1,3 @@
+const constraints = {
+  MAX_SMALL_STRING_LENGTH: 128,
+} as const;
