+++ runtime/lib/Handlers/function/runtime-command/trace/simple-trace-type.enum.ts
@@ -1,0 +1,10 @@
+export enum SimpleTraceType {
+  Text = 'text',
+  Speak = 'speak',
+  Audio = 'audio',
+  Debug = 'debug',
+  Visual = 'visual',
+  CardV2 = 'cardV2',
+  Carousel = 'carousel',
+  Choice = 'choice',
+}
